reorg, cli, improvements
* Finished reorganization of files/folders
* CLI logging & i18n is complete
* Better progress logging system
  * GUI language change will need to do some fancy stuff here
* DITTO mini flashing is fast enough now
  * README still needs update
* Some TODOs finished
