+++ pm2hw/linkers/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseLinker, linkers
+from .pokecard import PokeFlash
+from .dittomini import DittoFlash
