+++ pm2hw/locales/en/LC_MESSAGES/pm2hw.gui.po
@@ -149,14 +149,27 @@ msgstr "Flash the given file or use - to read from stdin."
 msgid "log.blocks.over"
 msgstr "Requested to access more than the available size, truncating request."
 
-msgid "log.dump.progress"
-msgstr "Dumping from {card.name} to {fn}"
-
-msgid "log.erase.progress"
-msgstr "Erasing data on {card.name}"
-
-msgid "log.flash.progress"
-msgstr "Flashing to {card.name} from {fn}"
+msgid "log.progress"
+msgstr "[DEFAULT]\n"
+"s=\n"
+"count=1\n"
+"gui={0}: {suffix}\n"
+"gui.initial=${gui}\n"
+"gui.medial=${gui}\n"
+"gui.final={0}: {suffix}\n ${s}  {completed}\n"
+"cli.initial=${s}  {0}\n"
+"cli.medial=${s}    ...{suffix}\n"
+"cli.final=${s}    {completed}\n"
+"suffix={cur}/{end} ({pc:.0f}%)\n"
+"completed=Completed in {secs:.3f}s\n"
+"[dump]\n"
+"message.0=Dumping from {card.name} to {fn}\n"
+"[erase]\n"
+"message.0=Erasing data on {card.name}\n"
+"[flash]\n"
+"message.0=Flashing to {card.name} from {fn}\n"
+"[verify]\n"
+"message.0=Verifying contents of {card.name}\n"
 
 msgid "log.ftdi.characters.disable"
 msgstr "* set EventChar=disabled ErrorChar=disabled"
@@ -181,9 +194,6 @@ msgstr "* set RequestTransferSize In={in} Out={out} (bytes)"
 
 msgid "log.ftdi.transfer.timeout.set"
 msgstr "* set TransferTimeout Read={read} Write={write} (ms)"
-
-msgid "log.verify.progress"
-msgstr "Verifying contents of {card.name}"
 
 msgid "log.verify.failed"
 msgstr "Verification failed"
@@ -391,9 +401,8 @@ msgstr "# <str name=help.topic.basics.title/>\n"
 "* Official flashing tool is Ditto Flash.\n"
 "* Must be used in a case, board will not make contact loose.\n\n"
 "#### Support in pm2hw\n"
-"Status: OK\n\n"
-"Supports basic operations (read, write, erase), but is currently slower"
-" than it should be.\n\n"
+"Status: Good\n\n"
+"Supports basic operations (read, write, erase).\n\n"
 "#### Technical specifications\n"
 "* Size: 2 MiB\n"
 "* Theoretical flashing speeds: 30s (512 KiB), 2m (2 MiB)\n"
@@ -945,3 +954,7 @@ msgstr "All files"
 
 msgid "misc.filetypes.min"
 msgstr "Pokemon mini ROMs"
+
+#: pm2hw\linkers\base.py:
+msgid "log.wait.cannot.buffer"
+msgstr "This device cannot buffer slow operations and may cause delays"
